+++ src/app/api/sign-in/route.ts
@@ -1,4 +1,4 @@
-import { signIn } from "@/common/auth/auth";
+import { signIn } from "@/infrastructures/auth/auth-provider";
 
 export async function GET() {
 	await signIn("auth0", { redirectTo: "/" }, { prompt: "login" });
