+++ src/middleware.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import createMiddleware from "next-intl/middleware";
-import { auth } from "@/common/auth/auth";
+import { auth } from "@/infrastructures/auth/auth-provider";
 import { routing } from "./i18n/routing";
 
 const handleI18nRouting = createMiddleware(routing);
